requirements.txt
# Web框架
streamlit>=1.28.0

# AI和机器学习
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.0.0

# HTTP请求
requests>=2.31.0

# 数据验证
pydantic>=2.0.0

# 环境变量管理
python-dotenv>=1.0.0

# 日期时间处理
python-dateutil>=2.8.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据可视化
plotly>=5.17.0

# 重试机制
retrying>=1.3.4

# 配置管理
pyyaml>=6.3.0

# 缓存
cachetools>=5.3.0

# JSON处理
orjson>=3.9.0